+++ tests/Support/Objects/ExamplesAuthorizeMessageQuery.php
@@ -1,0 +1,53 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rebing\GraphQL\Tests\Support\Objects;
+
+use Closure;
+use GraphQL\Type\Definition\ResolveInfo;
+use GraphQL\Type\Definition\Type;
+use Rebing\GraphQL\Support\Facades\GraphQL;
+use Rebing\GraphQL\Support\Query;
+
+class ExamplesAuthorizeMessageQuery extends Query
+{
+    protected $attributes = [
+        'name' => 'Examples authorize query',
+    ];
+
+    public function authorize($root, array $args, $ctx, ResolveInfo $resolveInfo = null, Closure $getSelectFields = null): bool
+    {
+        return false;
+    }
+
+    public function getAuthorizationMessage(): string
+    {
+        return 'You are not authorized to perform this action';
+    }
+
+    public function type(): Type
+    {
+        return Type::listOf(GraphQL::type('Example'));
+    }
+
+    public function args(): array
+    {
+        return [
+            'index' => ['name' => 'index', 'type' => Type::int()],
+        ];
+    }
+
+    public function resolve($root, $args, $context, ResolveInfo $resolveInfo, Closure $getSelectFields)
+    {
+        $data = include __DIR__.'/data.php';
+
+        if (isset($args['index'])) {
+            return [
+                $data[$args['index']],
+            ];
+        }
+
+        return $data;
+    }
+}